Okolab: VC++ warnings (integer type).
(IP port numbers are unsigned 16-bit.)

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12371 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Okolab/Okolab.cpp
@@ -191,7 +191,7 @@ void OkolabDevice::SetDeviceId(int id)
 }
 
 
-bool OkolabDevice::OCSConnectionOpen(int PortNo, char* IPAddress)
+bool OkolabDevice::OCSConnectionOpen(unsigned short PortNo, char* IPAddress)
 {
  int err;
  SOCKADDR_IN target;
